Pull BadRequest into its own file
Better separation of concerns for files makes them more legible and
discoverable.
